feat(restaurantController.js): add addRestaurant function to handle creation of new restaurants
The `addRestaurant` function is added to the `restaurantController.js` file. This function is responsible for creating a new restaurant instance based on the request body and saving it to the database. If the restaurant is successfully created, a response with status 201 and a success message is sent. If there is an error during the creation process, a response with status 500 and an error message is sent.