feat(categoriesController.js): add getAllCategory function to retrieve all categories from the database
The getAllCategory function is added to the categoriesController.js file. This function retrieves all categories from the database and returns them as a JSON response. If no categories are found, a 404 status code is returned. If there is an error during the retrieval process, a 500 status code is returned along with an error message.