feat(categoriesController.js): add getRandomCategory function to retrieve a random selection of categories
The getRandomCategory function is added to the categoriesController.js file. This function uses the aggregate method to query the Category collection and retrieve a random selection of categories. The query excludes the category with the value 'more' and returns a sample of 7 categories. If a category with the value 'more' exists, it is also included in the response. The function returns the selected categories in the response body with a status code of 200. If an error occurs during the execution of the function, an error message is returned in the response body with a status code of 500.