feat(foodController.js): add addFood function to handle adding a new food item
The addFood function is added to the foodController.js file. This function is responsible for creating a new instance of the Food model and saving it to the database. If the save operation is successful, a response with status 201 and a success message is sent. If there is an error during the save operation, a response with status 500 and the error message is sent.