feat(foodController.js): add getFoodById method to retrieve a specific food item by its ID
The getFoodById method is added to the foodController.js file. This method takes in a request parameter for the food ID and retrieves the corresponding food item from the database using the Food model. If the food item is found, it is returned with a 200 status code. If the food item is not found, a 404 status code is returned. If there is an error during the retrieval process, a 500 status code is returned along with an error message.