feat(foodController.js): add updateFoodById method to handle updating a food item by its ID
The updateFoodById method is added to the foodController.js file. This method takes in a request and response object and updates a food item in the database based on the provided ID.

If the food item is successfully updated, a 200 status code is returned along with a success message and the updated food item. If the food item is not found, a 404 status code is returned with an error message. If there is an error during the update process, a 500 status code is returned with an error message.