feat(driverController.js): add registerDriver function to handle driver registration
The registerDriver function is responsible for creating a new driver record in the database. It takes in a request object containing the driver details and saves it to the database using the Driver model. If the save operation is successful, it returns a 201 status code with a success message. If there is an error during the save operation, it returns a 500 status code with an error message.